+++ modules/kubernetes/aks-core/namespace.tf
@@ -11,6 +11,10 @@ resource "kubernetes_namespace" "tenant" {
     )
     name = each.value.name
   }
+
+  lifecycle {
+    prevent_destroy = true
+  }
 }
 
 resource "kubernetes_service_account_v1" "tenant" {